+++ database/migrations/2020_07_23_111237_create_candidate_experiences_table.php
@@ -13,36 +13,37 @@ class CreateCandidateExperiencesTable extends Migration
      */
     public function up()
     {
-        Schema::create('candidate_experiences', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('candidate_id');
-            $table->string('experience_title');
-            $table->text('company');
-            $table->unsignedBigInteger('country_id')->nullable();
-            $table->unsignedBigInteger('state_id')->nullable();
-            $table->unsignedBigInteger('city_id')->nullable();
-            $table->date('start_date');
-            $table->date('end_date')->nullable();
-            $table->boolean('currently_working')->default(0);
-            $table->text('description')->nullable();
-            $table->timestamps();
+        if(!Schema::hasTable("candidate_experiences"))
+            Schema::create('candidate_experiences', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('candidate_id');
+                $table->string('experience_title');
+                $table->text('company');
+                $table->unsignedBigInteger('country_id')->nullable();
+                $table->unsignedBigInteger('state_id')->nullable();
+                $table->unsignedBigInteger('city_id')->nullable();
+                $table->date('start_date');
+                $table->date('end_date')->nullable();
+                $table->boolean('currently_working')->default(0);
+                $table->text('description')->nullable();
+                $table->timestamps();
 
-            $table->foreign('candidate_id')->references('id')->on('candidates')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('candidate_id')->references('id')->on('candidates')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->foreign('country_id')->references('id')->on('countries')
-                ->onDelete('set null')
-                ->onUpdate('cascade');
+                $table->foreign('country_id')->references('id')->on('countries')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
 
-            $table->foreign('state_id')->references('id')->on('states')
-                ->onDelete('set null')
-                ->onUpdate('cascade');
+                $table->foreign('state_id')->references('id')->on('states')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
 
-            $table->foreign('city_id')->references('id')->on('cities')
-                ->onDelete('set null')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('city_id')->references('id')->on('cities')
+                    ->onDelete('set null')
+                    ->onUpdate('cascade');
+            });
     }
 
     /**
