+++ database/migrations/2020_07_23_111414_create_email_jobs_table.php
@@ -13,23 +13,24 @@ class CreateEmailJobsTable extends Migration
      */
     public function up()
     {
-        Schema::create('email_jobs', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('job_id');
-            $table->string('job_url');
-            $table->string('friend_name');
-            $table->string('friend_email');
-            $table->timestamps();
+        if(!Schema::hasTable("email_jobs"))
+            Schema::create('email_jobs', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedInteger('job_id');
+                $table->string('job_url');
+                $table->string('friend_name');
+                $table->string('friend_email');
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
+                $table->foreign('user_id')->references('id')->on('users')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
 
-            $table->foreign('job_id')->references('id')->on('jobs')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('job_id')->references('id')->on('jobs')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+            });
     }
 
     /**
