+++ database/migrations/2020_07_29_115041_create_candidate_language_table.php
@@ -13,20 +13,21 @@ class CreateCandidateLanguageTable extends Migration
      */
     public function up()
     {
-        Schema::create('candidate_language', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedInteger('language_id');
-            $table->timestamps();
+        if(!Schema::hasTable("candidate_language"))
+            Schema::create('candidate_language', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedInteger('language_id');
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table->foreign('language_id')->references('id')->on('languages')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table->foreign('language_id')->references('id')->on('languages')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-        });
+            });
     }
 
     /**
