+++ database/migrations/2020_08_14_115324_remove_social_links_from_companies.php
@@ -15,11 +15,16 @@ class RemoveSocialLinksFromCompanies extends Migration
     {
         
         Schema::table('companies', function (Blueprint $table) {
-            $table->dropColumn('facebook_url');
-            $table->dropColumn('twitter_url');
-            $table->dropColumn('linkedin_url');
-            $table->dropColumn('google_plus_url');
-            $table->dropColumn('pinterest_url');
+            if(Schema::hasColumn('companies', 'facebook_url'))
+                $table->dropColumn('facebook_url');
+            if(Schema::hasColumn('companies', 'twitter_url'))
+                $table->dropColumn('twitter_url');
+            if(Schema::hasColumn('companies', 'linkedin_url'))
+                $table->dropColumn('linkedin_url');
+            if(Schema::hasColumn('companies', 'google_plus_url'))
+                $table->dropColumn('google_plus_url');
+            if(Schema::hasColumn('companies', 'pinterest_url'))
+                $table->dropColumn('pinterest_url');
         });
     }
