+++ database/migrations/2020_10_01_101638_create_featured_records_table.php
@@ -13,22 +13,23 @@ class CreateFeaturedRecordsTable extends Migration
      */
     public function up()
     {
-        Schema::create('featured_records', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('owner_id');
-            $table->string('owner_type');
-            $table->unsignedBigInteger('user_id');
-            $table->string('stripe_id');
-            $table->dateTime('start_time');
-            $table->dateTime('end_time');
-            $table->text('meta');
-            $table->timestamps();
+        if(!Schema::hasTable("featured_records"))
+            Schema::create('featured_records', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedInteger('owner_id');
+                $table->string('owner_type');
+                $table->unsignedBigInteger('user_id');
+                $table->string('stripe_id');
+                $table->dateTime('start_time');
+                $table->dateTime('end_time');
+                $table->text('meta');
+                $table->timestamps();
 
-            $table->foreign('user_id')
-                ->references('id')->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+                $table->foreign('user_id')
+                    ->references('id')->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
     }
 
     /**
