+++ database/migrations/2020_10_10_053314_drop_featured_column_from_companies_and_jobs.php
@@ -13,13 +13,15 @@ class DropFeaturedColumnFromCompaniesAndJobs extends Migration
      */
     public function up()
     {
-        Schema::table('jobs', function (Blueprint $table) {
-            $table->dropColumn('is_featured');
-        });
+        if(Schema::hasColumn('jobs', 'is_featured'))
+            Schema::table('jobs', function (Blueprint $table) {
+                $table->dropColumn('is_featured');
+            });
 
-        Schema::table('companies', function (Blueprint $table) {
-            $table->dropColumn('is_featured');
-        });
+        if(Schema::hasColumn('companies', 'is_featured'))
+            Schema::table('companies', function (Blueprint $table) {
+                $table->dropColumn('is_featured');
+            });
     }
 
     /**
