+++ database/migrations/2021_04_20_055500_create_job_functional_areas_table.php
@@ -13,13 +13,14 @@ class CreateJobFunctionalAreasTable extends Migration
      */
     public function up()
     {
-        Schema::create('job_functional_areas', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('job_id');
-            $table->integer('functional_area_id');
-            $table->foreign('job_id')->references('id')->on('jobs')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable("job_functional_areas"))
+            Schema::create('job_functional_areas', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedInteger('job_id');
+                $table->integer('functional_area_id');
+                $table->foreign('job_id')->references('id')->on('jobs')->onDelete('cascade');
+                $table->timestamps();
+            });
     }
 
     /**
